+++ LineSharp/Messages/SendMessageObject.cs
@@ -126,10 +126,10 @@ namespace LineSharp.Messages
     public class Size
     {
         [JsonProperty("width")]
-        public int Width { get; set; } = 1040;
+        public double Width { get; set; } = 1040;
 
         [JsonProperty("height")]
-        public int Height { get; set; }
+        public double Height { get; set; }
     }
 
     [JsonConverter(typeof(JsonSubtypes))]
@@ -163,16 +163,16 @@ namespace LineSharp.Messages
     public class ImagemapAreaObject
     {
         [JsonProperty("x")]
-        public int X { get; set; }
+        public double X { get; set; }
 
         [JsonProperty("y")]
-        public int Y { get; set; }
+        public double Y { get; set; }
 
         [JsonProperty("width")]
-        public int Width { get; set; }
+        public double Width { get; set; }
 
         [JsonProperty("height")]
-        public int Height { get; set; }
+        public double Height { get; set; }
     }
 
     [JsonSubtype(MessageType.Template)]
